Rename HsBindgen.Frontend.Macros.XYZ to HsBindgen.Frontend.Macro.XZY
Preparation for moving other Macro modules to Frontend.
