+++ hs-bindgen/src-internal/HsBindgen/C/Reparse/Macro.hs
@@ -9,15 +9,16 @@ import Text.Parsec
 import Text.Parsec.Expr
 
 import Clang.LowLevel.Core
-import HsBindgen.C.Reparse.Common ( reparseName, reparseLocName )
+import HsBindgen.C.Reparse.Common (reparseLocName, reparseName)
 import HsBindgen.C.Reparse.Infra
 import HsBindgen.C.Reparse.Literal
 import HsBindgen.C.Tc.Macro qualified as Macro
 import HsBindgen.Frontend.AST.External qualified as C
-import HsBindgen.Frontend.Macros.AST.Syntax
+import HsBindgen.Frontend.Macro.AST.Syntax
 import HsBindgen.Language.C qualified as C
 
-import {-# SOURCE #-} HsBindgen.C.Reparse.Decl ( reparseTypeName, reparseAttributeSpecifier )
+import {-# SOURCE #-} HsBindgen.C.Reparse.Decl (reparseAttributeSpecifier,
+                                                reparseTypeName)
 
 {-------------------------------------------------------------------------------
   Top-level
