+++ hs-bindgen/src-internal/HsBindgen/C/Tc/Macro.hs
@@ -102,7 +102,7 @@ import C.Expr.HostPlatform qualified as C.Expr
 -- hs-bindgen
 import HsBindgen.C.Tc.Macro.Type
 import HsBindgen.Errors
-import HsBindgen.Frontend.Macros.AST.Syntax
+import HsBindgen.Frontend.Macro.AST.Syntax
 import HsBindgen.Frontend.AST.External qualified as C
 import HsBindgen.Imports
 import HsBindgen.Language.C qualified as C
